Started integration of the (m,n,k)-domain and Negamax
