+++ src/main/java/za/ac/sun/cs/adversarial/algorithm/Negamax.java
@@ -1,6 +1,6 @@
 package za.ac.sun.cs.adversarial.algorithm;
 
-import za.ac.sun.cs.adversarial.domain.MNKGame;
+import za.ac.sun.cs.adversarial.domain.Board;
 import za.ac.sun.cs.adversarial.domain.Move;
 
 import java.util.List;
@@ -18,9 +18,9 @@ public class Negamax {
      * @param color
      * @return
      */
-    public static int F1(MNKGame node, int depth, int bound, int color) {
-        if (depth == 0 || node.isTerminal()) {
-            return color * node.value();
+    public static int F1(Board node, int depth, int bound, int color) {
+        if ((depth == 0) || node.isTerminal() < 0) {
+            return color;
         }
 
         List<Move> moves= node.getLegalMoves();
@@ -28,7 +28,7 @@ public class Negamax {
         int value = Integer.MIN_VALUE;
 
         for (Move move : moves) {
-            node.makeMove(move);
+            node.makeMove(color, move);
 
             value = max(value, -F1(node, depth - 1, -bound, -color));
 
@@ -52,9 +52,9 @@ public class Negamax {
      * @param color
      * @return
      */
-    public static int F2(MNKGame node, int depth, int alpha, int beta, int color) {
-        if (depth == 0 || node.isTerminal()) {
-            return color * node.value();
+    public static int F2(Board node, int depth, int alpha, int beta, int color) {
+        if (depth == 0 || node.isTerminal() < 0) {
+            return color;
         }
 
         List<Move> moves= node.getLegalMoves();
@@ -62,7 +62,7 @@ public class Negamax {
         int value = Integer.MIN_VALUE;
 
         for (Move move : moves) {
-            node.makeMove(move);
+            node.makeMove(color, move);
 
             value = max(value, -F2(node, depth - 1, -beta, -alpha, -color));
