+++ src/main/java/za/ac/sun/cs/adversarial/domain/Board.java
@@ -1,9 +1,7 @@
-package za.ac.sun.cs.search.adversarialagent;
+package za.ac.sun.cs.adversarial.domain;
 
 import java.util.ArrayList;
 import java.lang.StringBuilder;
-
-import za.ac.sun.cs.search.adversarialagent.Move;
 
 public class Board {
     private final int m;
@@ -12,7 +10,7 @@ public class Board {
     private final int[][] board;
     private int player;
 
-    public Board (int m, int n, int k) {
+    public Board(int m, int n, int k) {
         this.m = m;
         this.n = n;
         this.k = k;
@@ -20,7 +18,7 @@ public class Board {
         this.player = 1;
     }
 
-    public ArrayList<Move> getLegalMoves () {
+    public ArrayList<Move> getLegalMoves() {
         ArrayList<Move> moves = new ArrayList<Move>();
         for (int i = 0; i < this.m; i++) {
             for (int j = 0; j < this.n; j++) {
@@ -33,7 +31,12 @@ public class Board {
         return moves;
     }
 
-    public void makeMove (int player, int i, int j) {
+
+    public void makeMove(int player, Move move) {
+        this.board[move.getRow()][move.getColumn()] = player;
+    }
+
+    public void makeMove(int player, int i, int j) {
         this.board[i][j] = player;
     }
 
@@ -41,7 +44,7 @@ public class Board {
 
         // Check Row wins
         for (int i = 0; i <= this.n - this.k; i++) {
-            for (int j = 0; j < this.m; j++) {    
+            for (int j = 0; j < this.m; j++) {
                 if (checkRowWin(1, j, i)) {
                     return 1;
                 } else if (checkRowWin(2, j, i)) {
@@ -53,7 +56,7 @@ public class Board {
         // Check Column wins
 
         for (int i = 0; i <= this.n - this.k; i++) {
-            for (int j = 0; j < this.n; j++) {    
+            for (int j = 0; j < this.n; j++) {
                 if (checkColumnWin(1, j, i)) {
                     return 1;
                 } else if (checkColumnWin(2, j, i)) {
@@ -64,20 +67,19 @@ public class Board {
 
         // Check Diagonal wins
         for (int i = 0; i <= this.n - this.k; i++) {
-            for (int j = 0; j <= this.m - this.k; j++) {    
+            for (int j = 0; j <= this.m - this.k; j++) {
                 if (checkDiagonalWin(1, j, i)) {
                     return 1;
                 } else if (checkDiagonalWin(2, j, i)) {
                     return 2;
                 }
-            }        
+            }
         }
 
         // Check for a draw.
-        if (getLegalMoves(1).size() == 0 && getLegalMoves(2).size() == 0) {
-            return 0; 
+        if (getLegalMoves().size() == 0) {
+            return 0;
         }
-
 
 
         // Nobody has won.
@@ -85,23 +87,23 @@ public class Board {
     }
 
     public boolean checkRowWin(int player, int row, int col) {
-            for (int j = col; j < col + this.k; j++) {
-                if (this.board[row][j] != player) {
-                    return false;
-                }
+        for (int j = col; j < col + this.k; j++) {
+            if (this.board[row][j] != player) {
+                return false;
             }
-    
+        }
+
 
         return true;
     }
 
     public boolean checkColumnWin(int player, int col, int row) {
-           for (int j = row; j < row + this.k; j++) {
-                if (this.board[j][col] != player) {
-                    return false;
-                }
+        for (int j = row; j < row + this.k; j++) {
+            if (this.board[j][col] != player) {
+                return false;
             }
-    
+        }
+
 
         return true;
     }
@@ -111,24 +113,24 @@ public class Board {
         boolean forward = true, reverse = true;
 
         // Check forward diagonal
-            for (int j = col; j < this.k; j++) {
-                if (this.board[row + j][j] != player) {
-                    forward = false;
-                    break;
-                }
-           
+        for (int j = col; j < this.k; j++) {
+            if (this.board[row + j][j] != player) {
+                forward = false;
+                break;
             }
+
+        }
 
         // Check backward diagonal
 
-            for (int j = col; j < this.k; j++) {
-                if (this.board[row + j][this.n - j - 1] != player) {
-                    reverse = false;
-                    break;
-                }
-           }        
+        for (int j = col; j < this.k; j++) {
+            if (this.board[row + j][this.n - j - 1] != player) {
+                reverse = false;
+                break;
+            }
+        }
 
-       return (forward || reverse);
+        return (forward || reverse);
     }
 
     public int getPlayer() {
@@ -151,7 +153,7 @@ public class Board {
 
         StringBuilder sb = new StringBuilder();
 
-        for (int i = 0 ; i < this.m; i++) {
+        for (int i = 0; i < this.m; i++) {
             for (int j = 0; j < this.n; j++) {
                 sb.append(board[i][j]);
             }
