--- src/main/java/za/ac/sun/cs/adversarial/domain/MNKGame.java
@@ -1,20 +1,0 @@
-package za.ac.sun.cs.adversarial.domain;
-
-import java.util.List;
-
-public class MNKGame {
-    public boolean isTerminal() {
-    }
-
-    public int value() {
-    }
-
-    public List<Move> getLegalMoves() {
-    }
-
-    public void makeMove(Move move) {
-    }
-
-    public void undoMove() {
-    }
-}
